mon: always forward with uniform priority
If a client sends us a high priority request, forward it to the leader mon
using a normal priority.  This avoids potential confusion with ordering of
messages between monitors.

This explains symptom observed in #4810, and may be the root issue if there
is starvation.

Note that this is partly problematic because ceph_mon.cc sets the default
send priority to HIGH, and the MonClients are using the NORMAL priority.
Making the mons use NORMAL is probably cleanest, but it might be awkward to
make that switch, and having everything at the same (HIGH) level should be
just fine.

Signed-off-by: Sage Weil <sage@inktank.com>