librbd: clear LAYERING feature bit when removing the parent
Once we remove the parent, clients no longer need to support layering to
use the image.  Remove the feature bit.

This also makes it simpler for clients to detect whether a parent is
present in that the get_parent ENOENT needn't be handled (although it
probably should be given that the old behavior did not clear this bit).

Signed-off-by: Sage Weil <sage@inktank.com>